pallet_reserve::pallet
Module dispatchables

Auto-generated docs-only module listing all defined dispatchables for this pallet.

Warning: Doc-Only

Members of this module cannot be used directly and are only provided for documentation purposes. To see the real version of each dispatchable, look for them in Pallet or Call.

Functions

apply_as

Dispatch a call as coming from the reserve account

spend

Spend amount funds from the reserve account to to.

tip

Deposit amount tokens in the treasure account