Struct pallet_reserve::pallet::Pallet

pub struct Pallet<T, I = ()>(_);

Implementations

impl<T: Config<I>, I: 'static> Pallet<T, I>

pub fn spend(
origin: OriginFor<T>,
to: T::AccountId,
amount: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Spend amount funds from the reserve account to to.

pub fn tip(
origin: OriginFor<T>,
amount: <<T as Config<I>>::Currency as Currency<<T as Config>::AccountId>>::Balance
) -> DispatchResultWithPostInfo

Deposit amount tokens in the treasure account

pub fn apply_as(
origin: OriginFor<T>,
call: Box<<T as Config<I>>::Call>
) -> DispatchResultWithPostInfo

Dispatch a call as coming from the reserve account